# Optimizing array mapping in the kernel.