To quote the commit message:

>  We unroll the map operation by hand in two performance-critical cases so as
not to call the generic array allocation function in OCaml, and allocate
directly on the minor heap instead. The generic array function is slow because
it needs to discriminate between float arrays and others, and furthermore it
jumps into the GC for pending operations. The unrolling replaces this by a simple
increment to the minor heap pointer and moves from the stack.
>
> The quantity of unrolling was determined by experimental measures on various Coq
developments. It looks like most of the maps are for small arrays of size lesser
or equal to 4, so this is what is implemented here. We could probably extend it
to an even bigger number, but that would result in ugly code. From what I've seen,
virtually all maps are of size less than 16, so that we could probably be almost
optimal by going up to 16 unrollings, but the code tradeoffs are not obvious. Maybe
when we have PPX?

I measured a 10% improvement in compilation time of UniMath (9.7% to be exact), where I first spotted the problem, although it is probably pervasive in developments doing computation in the kernel *for realz*.

I think it is innocuous enough to be integrated in 8.6, so that it will provide an important speedup for such developments. It does not change the API nor the observable behaviour of the kernel, except for speedup.