@ppedrot 2017-02-17
Note that in the particular case of UniMath, it does not solve anything about the kernel being stupid and gladly unfolding stuff until the end of the universe as we know it, so that it cannot be considered a solution to their slowness issues in general.