@aspiwack 2017-02-17
There seems to be two different optimisations here: loop unrolling and specialisation.

Have you compared with loop unrolling without specialisation? (I confess would feel more comfortable if this sort of code was in a generic library)