@ppedrot 2017-02-17
You can't really try. As soon as you don't have specialization, OCaml calls the C function `array_make`, which defeats the purpose. I don't know if this is a bug of the OCaml compiler, or a limitation of the boxed float representation.