@aspiwack 2017-02-17
Can you elaborate (for my own curious mind) on the different code-path taken by

```ocaml
[| v0 ; v1 ; v2 |]
```

Depending on whether is has a known type or not?