@ppedrot 2017-02-17
Consider:
```ocaml
let f_int (x : int) = [|x|]
let g_box (x : int option) = [|x|]
let h_any (x : 'a) = [|x|]
```
The corresponding assembly excerpt is:
```asm
# Code for f

camlFoo__f_int_1008:
	pushq	%rbp
	movq	%rsp, %rbp
	subq	$0, %rsp
	movq	%rax, %rbx
.L101:	subq	$16, %r15
	movq	caml_young_limit@GOTPCREL(%rip), %rax
	cmpq	(%rax), %r15
	jb	.L102
	leaq	8(%r15), %rax
	movq	$1024, -8(%rax)
	movq	%rbx, (%rax)
	addq	$0, %rsp
	popq	%rbp
	ret
.L102:	call	caml_call_gc@PLT
.L103:	jmp	.L101

# Code for g

camlFoo__g_box_1010:
	pushq	%rbp
	movq	%rsp, %rbp
	subq	$0, %rsp
	movq	%rax, %rbx
.L105:	subq	$16, %r15
	movq	caml_young_limit@GOTPCREL(%rip), %rax
	cmpq	(%rax), %r15
	jb	.L106
	leaq	8(%r15), %rax
	movq	$1024, -8(%rax)
	movq	%rbx, (%rax)
	addq	$0, %rsp
	popq	%rbp
	ret
.L106:	call	caml_call_gc@PLT
.L107:	jmp	.L105

# Code for h

camlFoo__h_any_1012:
	pushq	%rbp
	movq	%rsp, %rbp
	subq	$0, %rsp
.L108:
	movq	%rax, %rbx
.L109:	subq	$16, %r15
	movq	caml_young_limit@GOTPCREL(%rip), %rax
	cmpq	(%rax), %r15
	jb	.L110
	leaq	8(%r15), %rdi
	movq	$1024, -8(%rdi)
	movq	%rbx, (%rdi)
	movq	caml_make_array@GOTPCREL(%rip), %rax
	call	caml_c_call@PLT
	movq	caml_young_ptr@GOTPCREL(%rip), %r11
	movq    (%r11), %r15
	addq	$0, %rsp
	popq	%rbp
	ret
.L110:	call	caml_call_gc@PLT
.L111:	jmp	.L109
```

Note the call to `caml_make_array` in the polymorphic case.