@aspiwack 2017-02-17
I'm a little confused that it looks like `caml_make_array` is called after anything of interest happens.

Anyway, I gather you are only interested in the loop unrolling because it prevents a call to `caml_make_array`. The comment in the commit about the GC seems inaccurate though: all three function seem to behave the same in this respect (if pointer passes the limit then jump into the GC).