@ppedrot 2017-02-17
No, this is actually the `caml_make_array` function which is doing fancy things w.r.t. the GC. The comment is indeed a bit ambiguous on that matter.