@silene 2017-02-17
> No, this is actually the `caml_make_array` function which is doing fancy things w.r.t. the GC.

Is that right? As far as I remember, `caml_make_array` is just the identity function for anything that is not a floating-point array. In particular it does not call the GC in any way. By the way, Arnaud, that is why it seems to you that "`caml_make_array` is called after anything of interest happens." You are actually right, everything interesting happens before the call because the function is just the identity.