@ppedrot 2017-02-17
> Is that right? As far as I remember, caml_make_array is just the identity function for anything that is not a floating-point array.

Yes, but to check that, this function needs to call the `Is_in_value_area` macro, which in turns call `caml_page_table_lookup`. And this is where things start to be costly.