@ppedrot 2017-02-17
Right, that's not quite the GC, so the comment is wrong. What about just writing something along the lines of "not to call the `caml_array_make` function" in the comment?