@ppedrot 2017-02-19
I've updated the comment. Is there any arguments against merging this in 8.6? Can somebody like @maximedenes proceed to the merge if not?