@gares 2017-02-20
My impression on this class of patches is that we should not do what the compiler is supposed to do.
Even more when the compiler is able to do it:
  https://caml.inria.fr/pub/docs/manual-ocaml/flambda.html
see `-O3` and the `@unrolled n` attribute.