@ppedrot 2017-02-20
Indeed, I fully agree that in an ideal world OCaml should do this for us.

Problem is, as of today, it does not apply the unrolling, while the performance gain is quite noticeable. Furthermore, assuming we use Flambda I'm not even sure that it would unroll the Array.make function because it is primitive and written in C. And lastly, I don't think adding a gazillion of optimization annotations is functionally and philosophically better than carefully hand-written code (obviously that's not true for eldritch code, but here the ratio code effectiveness / ugliness seems reasonable).