@maximedenes 2017-02-21
Is it really 8.6 that you want this in? Optimizations typically go in the version currently in development, right?