@ppedrot 2017-02-21
I personally don't care because I'm living on trunk, but there are probably a handful of users out there that would bless the fact Coq got quite quicker in between pl. For a trivial patch that is virtually not observable in the source code, I think this does not matter whether it goes on the latest release or in trunk. (Plus how does 8d783c1 qualify? I know it was officially about *reenabling* an optim, but the code had been rewritten, so that in practice it was *implementing* an optim.)