@gares 2017-02-22
8d783c1 is fixing a regression.  This PR is an improvement over any released Coq version.