@ppedrot 2017-02-22
Fine, fine, I can reopen a PR against trunk if there is general consensus. Should I?