@aspiwack 2017-02-22
You can just use the « edit » button near the PR title to change the branch it targets. Probably better.