@ejgallego 2017-02-23
So assume we are talking about a 5% faster Coq, times 8 months until 8.7 is out, times the number of Coq users, it certainly looks like an awful number of hours gained to me, the benefit is clear.

On the other hand, the what are the downsides/dangers of merging this 8.6? None that I can see, the patch is trivial enough. So IMHO, this seems like a clear 8.6 candidate.