@ejgallego 2017-02-23
> Optimizations typically go in the version currently in development, right?

Why? I'd say _invasive_ optimizations indeed, but I think few people would reject a trivial patch that gives you a non-trivial speed gain, specially when the next stable version is 8 months away.