@ppedrot 2017-02-23
Too add a little spice to this conversion that is slowly but surely turning into a brawl contest, we could probably ask people concerned by this optimization. I'd for instance be interested in hearing from @mortberg (from now on my official UniMath contact) and @JasonGross (I'm pretty sure all the crazy stuff he writes in fiat is heavily affected by this patch, by judging the quantity of Defined back there).