@maximedenes 2017-02-23
I'd like to compensate the spice and relax the atmosphere by saying that we can discuss a more flexible policy for the integration of optimizations :) It is true that with a 10-month release cycle, it is a pity to postpone a trivial patch. On the other hand, how do we draw the line between "invasive" optimizations and innocuous ones?

We can also choose to discuss this particular patch at the next WG, independently from any policy.

In any case, we should also report the problem to the OCaml team, to see if they have plans to improve the compiler in such scenarios.