@matejkosik 2017-03-24
Maxime asked me to run benchmark for this branch. [Here](https://ci.inria.fr/coq/view/benchmarking/job/benchmark-the-whole-branch/72/consoleText) are the results:
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │           CPU instructions            │
│                        │                         │                                     │                                       │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│             coq-geocoq │ 2215.65 2394.71 -7.48 % │ 6170994808068 6656242409157 -7.29 % │ 10361347647765 11375764063691 -8.92 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│     coq-mathcomp-field │  448.16  482.09 -7.04 % │ 1247302767034 1342347020178 -7.08 % │  2102120885665  2265682765580 -7.22 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│            coq-unimath │ 1210.59 1286.37 -5.89 % │ 3355432194448 3568318753607 -5.97 % │  5603408474428  5966451346906 -6.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-character │  257.09  266.82 -3.65 % │  714931358748  742477217286 -3.71 % │  1059788621679  1106919778090 -4.26 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-fingroup │   56.05   58.10 -3.53 % │  154738004192  159982326167 -3.28 % │   216940823047   225441075129 -3.77 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-solvable │  193.86  200.76 -3.44 % │  538922552219  556319616458 -3.13 % │   799174021367   831012275426 -3.83 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│   coq-mathcomp-algebra │  167.05  171.74 -2.73 % │  463162625899  476465657132 -2.79 % │   658656208840   682405638449 -3.48 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-iris │  425.85  430.56 -1.09 % │ 1183018356474 1192535580136 -0.80 % │  1714537671015  1723851168800 -0.54 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│           coq-compcert │  818.24  826.58 -1.01 % │ 2267292209274 2293076056489 -1.12 % │  3477884405878  3498242752193 -0.58 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   42.99   43.26 -0.62 % │  117760794283  118577057612 -0.69 % │   150366159418   151805550220 -0.95 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│       coq-fiat-parsers │  511.12  513.78 -0.52 % │ 1402393639585 1409121212222 -0.48 % │  2001361491386  2024058480020 -1.12 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-corn │ 1531.37 1538.66 -0.47 % │ 4253260785911 4269231710019 -0.37 % │  6780119842121  6849057101131 -1.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│        coq-fiat-crypto │ 2305.33 2315.82 -0.45 % │ 6410038018359 6439959679069 -0.46 % │ 10816514719325 10864689805515 -0.44 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-flocq │   51.68   51.80 -0.23 % │  142738939564  143166801865 -0.30 % │   182462153078   183725483337 -0.69 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-color │  632.36  632.66 -0.05 % │ 1752403745130 1752253873035 +0.01 % │  2279404804327  2289911626709 -0.46 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│         coq-coquelicot │   70.09   70.10 -0.01 % │  193288158782  194036980407 -0.39 % │   242828869250   244756368150 -0.79 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                 coq-sf │   44.31   44.14 +0.39 % │  120895539792  120965434115 -0.06 % │   158704403417   159864692564 -0.73 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│       coq-math-classes │  222.73  220.96 +0.80 % │  612742581504  608562872886 +0.69 % │   859633068673   862349723029 -0.32 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-hott │  230.22  228.16 +0.90 % │  621766384982  616291778379 +0.89 % │  1039452768055  1043443935023 -0.38 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 2b8ad7e04002ebe9fec5790da924673418f2fa7f
OLD = 7707396c5010d88c3d0be6ecee816d8da7ed0ee0
```
The table shows how much time does it take to compile the selected OPAM packags.
- at the bottom of this branch (where it was fored from `trunk`) ... this commit is denoted as `OLD`
- at the top of this branch ... this commit is denoted as `NEW`

So e.g. in case of `coq-geocoq` we see that the compilation time is shortened cca. 7%.